Add stock retrieval endpoints and update data models for improved querying
- Implemented new endpoints in Stock3Controller for fetching stocks by symbol, company name, price range, top prices, and pagination.
- Updated Stock3Dto to use decimal for price instead of double for better precision.
- Enhanced IStock3Repository and IStock3Service interfaces with new methods for additional stock queries.
- Added PagedResult class for handling paginated responses.

+++ API/DTOs/Stock3Dto.cs
@@ -11,7 +11,7 @@ namespace API.DTOs
         public string Symbol { get; set; }
         public string CompanyName { get; set; }
         public string Industry { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public long MarketCap { get; set; }
         public DateTime LastUpdated { get; set; }
     }
@@ -20,7 +20,7 @@ namespace API.DTOs
         public string Symbol { get; set; }
         public string CompanyName { get; set; }
         public string Industry { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public long MarketCap { get; set; }
 
     }
@@ -29,7 +29,7 @@ namespace API.DTOs
         public string? Symbol { get; set; }
         public string? CompanyName { get; set; }
         public string? Industry { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         public long? MarketCap { get; set; }
      
     }
